# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Perimtr API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  /** @nullable */
  attemptsRemaining?: number | null;
}

export interface DashboardStats {
  totalSessions: number;
  totalAttendees: number;
  totalPolls: number;
  totalVotes: number;
  activeSessions: number;
  activePolls: number;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  active: 'active',
  expired: 'expired',
  ended: 'ended',
} as const;

export interface Session {
  id: number;
  title: string;
  pin: string;
  status: SessionStatus;
  durationSeconds: number;
  networkKey: string;
  /** @nullable */
  attendeeCount?: number | null;
  createdAt: string;
  expiresAt: string;
}

export interface SessionInput {
  /** @minLength 1 */
  title: string;
  /** @minimum 30 */
  durationSeconds: number;
}

export type SessionUpdateStatus = typeof SessionUpdateStatus[keyof typeof SessionUpdateStatus];


export const SessionUpdateStatus = {
  ended: 'ended',
} as const;

export interface SessionUpdate {
  status?: SessionUpdateStatus;
}

export interface TimePoint {
  time: string;
  count: number;
}

export interface SessionStats {
  sessionId: number;
  title: string;
  attendeeCount: number;
  status: string;
  expiresAt: string;
  checkInsOverTime?: TimePoint[];
}

export interface Attendee {
  id: number;
  sessionId: number;
  name: string;
  deviceToken: string;
  checkedInAt: string;
  /** @nullable */
  ipAddress?: string | null;
}

export interface CheckInInput {
  /** @minLength 1 */
  name: string;
  /**
     * @minLength 4
     * @maxLength 4
     */
  pin: string;
  deviceToken: string;
}

export interface CheckInResult {
  success: boolean;
  attendeeId: number;
  name: string;
}

export type PollStatus = typeof PollStatus[keyof typeof PollStatus];


export const PollStatus = {
  active: 'active',
  ended: 'ended',
} as const;

export interface Poll {
  id: number;
  question: string;
  options: string[];
  /** @nullable */
  pin?: string | null;
  pinRequired: boolean;
  networkKey?: string;
  status: PollStatus;
  /** @nullable */
  totalVotes?: number | null;
  createdAt: string;
}

export interface PollInput {
  /** @minLength 1 */
  question: string;
  /** @minItems 2 */
  options: string[];
  pinRequired: boolean;
}

export type PollUpdateStatus = typeof PollUpdateStatus[keyof typeof PollUpdateStatus];


export const PollUpdateStatus = {
  ended: 'ended',
} as const;

export interface PollUpdate {
  status?: PollUpdateStatus;
  pinRequired?: boolean;
}

export interface VoteInput {
  /** @minimum 0 */
  optionIndex: number;
  deviceToken: string;
  /** @nullable */
  pin?: string | null;
}

export interface VoteResult {
  success: boolean;
  optionIndex: number;
}

export interface PollOption {
  label: string;
  votes: number;
  percentage: number;
}

export interface PollResults {
  pollId: number;
  question: string;
  options: PollOption[];
  totalVotes: number;
  status?: string;
}

export interface DeviceFingerprint {
  fingerprint: string;
  /** @nullable */
  userAgent?: string | null;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  screenResolution?: string | null;
  /** @nullable */
  timezone?: string | null;
  /** @nullable */
  hardwareConcurrency?: number | null;
  /** @nullable */
  deviceMemory?: string | null;
}

export interface DeviceTokenResponse {
  token: string;
  isNew: boolean;
}

export interface NetworkVerifyInput {
  /** @nullable */
  sessionId?: number | null;
  /** @nullable */
  pollId?: number | null;
}

export interface NetworkVerifyResult {
  allowed: boolean;
  reason: string;
}

export interface AuthUser {
  id: string;
  email: string;
  name: string;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type LogoutAdmin200 = {
  success: boolean;
};