# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Perimtr API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List all attendance sessions
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new attendance session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
...
  /auth/register:
    post:
      tags: [auth]
      operationId: registerAdmin
      summary: Create a new admin account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: Account created and session started.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '409':
          description: Email already in use.
  /auth/login:
    post:
      tags: [auth]
      operationId: loginAdmin
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Signed in successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: Invalid credentials.
...
components:
...
  schemas:
...
    Session:
      type: object
      required: [id, title, pin, status, durationSeconds, networkKey, createdAt, expiresAt]
      properties:
        id:
          type: integer
        title:
          type: string
        pin:
          type: string
        status:
          type: string
          enum: [active, expired, ended]
        durationSeconds:
          type: integer
        networkKey:
          type: string
        attendeeCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        expiresAt:
          type: string
    SessionInput:
      type: object
      required: [title, durationSeconds]
      properties:
        title:
          type: string
          minLength: 1
        durationSeconds:
          type: integer
          minimum: 30
...
    AuthUser:
      type: object
      required: [id, email, name]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1